+++ src/error.rs
@@ -2,6 +2,42 @@ use geozero::error::GeozeroError;
 use h3o::error::OutlinerError;
 use std::{error::Error, fmt};
 
+/// Error occurring while rendering a set of cell indices to MVT.
+#[derive(Debug)]
+#[non_exhaustive]
+pub enum InvalidTileID {
+    /// Invalid X coordinate.
+    InvalidX(u32),
+    /// Invalid Y coordinate.
+    InvalidY(u32),
+    /// Invalid Z coordinate.
+    InvalidZ(u32),
+}
+
+impl fmt::Display for InvalidTileID {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match *self {
+            Self::InvalidX(value) => {
+                write!(f, "invalid x coordinate: {value}")
+            }
+            Self::InvalidY(value) => {
+                write!(f, "invalid y coordinate: {value}")
+            }
+            Self::InvalidZ(value) => {
+                write!(f, "invalid z coordinate: {value}")
+            }
+        }
+    }
+}
+
+impl Error for InvalidTileID {
+    fn source(&self) -> Option<&(dyn Error + 'static)> {
+        match *self {
+            Self::InvalidX(_) | Self::InvalidY(_) | Self::InvalidZ(_) => None,
+        }
+    }
+}
+
 /// Errors occurring while rendering a set of cell indices to MVT.
 #[derive(Debug)]
 #[non_exhaustive]
@@ -47,6 +83,9 @@ mod tests {
         assert!(!RenderingError::Encoding(GeozeroError::GeometryFormat)
             .to_string()
             .is_empty());
+        assert!(!InvalidTileID::InvalidX(42).to_string().is_empty());
+        assert!(!InvalidTileID::InvalidY(42).to_string().is_empty());
+        assert!(!InvalidTileID::InvalidZ(42).to_string().is_empty());
     }
 
     #[test]
@@ -59,5 +98,8 @@ mod tests {
         assert!(RenderingError::Encoding(GeozeroError::GeometryFormat)
             .source()
             .is_some());
+        assert!(InvalidTileID::InvalidX(42).source().is_none());
+        assert!(InvalidTileID::InvalidY(42).source().is_none());
+        assert!(InvalidTileID::InvalidZ(42).source().is_none());
     }
 }